+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renamed ImageButton to ChapterButton
 - Renamed ItemsPanel to ChapterItemsPanel
 - Renamed NumberBox to ChapterNumberBox
+- Renamed OptionButton to ChapterToggleSwitch
 ### Supported .Net Versions
 - .Net Core 3.0
 - .Net Framework 4.5